refactoring UX
